+++ src/main/kotlin/com/redblue/web/company/model/dto/StockExcelDto.kt
@@ -7,27 +7,27 @@ data class StockExcelDto(
 
 	val companyName: String? = null,
 
-	val amount: Int = 0,
+	val amount: Long = 0,
 
 	val amountUpdatedAt: Date? = null,
 
-	val scheduleCount: Int = 0,
+	val scheduleCount: Long = 0,
 
 	val scheduleCountUpdatedAt: Date? = null,
 
-	val issuedCount: Int = 0,
+	val issuedCount: Long = 0,
 
 	val issuedCountUpdatedAt: Date? = null,
 
-	val normalCount: Int = 0,
+	val normalCount: Long = 0,
 
-	val firstCount: Int = 0,
+	val firstCount: Long = 0,
 
-	val noFaceValueCount: Int = 0,
+	val noFaceValueCount: Long = 0,
 
 	val noFaceValueUpdatedAt: Date? = null,
 
-	val noFaceValueCapitalAmount: Int = 0,
+	val noFaceValueCapitalAmount: Long = 0,
 
 	val noFaceValueCapitalAmountUpdatedAt: Date? = null
